Create a toplevel class for the Metadata
